# packages/contracts/openapi/visits.yaml
openapi: "3.1.0"
info:
  title: Smart VMS - Visitor & Visits API (US-11)
  version: "1.0.0"
  description: >
    Portal (unauthenticated) visit-request submission and authenticated
    host/reception visit endpoints. Reused verbatim from ADR-001's
    tenancy/RLS pattern; the 403 (tenant scope) and 409 (invalid transition)
    response declarations below are mandatory per .claude/rules/contracts.md.
paths:
  /public/portal/{tenant_slug}/visit-requests:
    post:
      summary: Submit a visitor pre-registration (public, unauthenticated)
      operationId: submitPortalVisitRequest
      parameters:
        - name: tenant_slug
          in: path
          required: true
          schema: { type: string }
        - name: Idempotency-Key
          in: header
          required: false
          description: >
            Optional. Dedup is tenant-scoped and content-hashed (a hash of
            contact_value + host_hint), NOT the raw header value -- a
            guessed/reused key can never return an unrelated submission's
            tracking reference.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalVisitRequestCreate"
      responses:
        "202":
          description: >
            Accepted. Uniform response regardless of whether host_hint
            resolved to a known host (anti-enumeration) -- never host or
            other-visit data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalVisitRequestAccepted"
        "400":
          description: CAPTCHA verification failed. No DB write occurred.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: >
            Unknown or suspended tenant_slug. Identical status+detail to
            every other 404 this endpoint can produce (anti-enumeration).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: privacy_notice_acknowledged missing or false.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Per-IP or per-tenant-slug rate limit exceeded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  schemas:
    PortalVisitRequestCreate:
      type: object
      required:
        - visitor_full_name
        - contact_channel
        - contact_value
        - privacy_notice_acknowledged
        - privacy_notice_version
        - turnstile_token
      properties:
        visitor_full_name: { type: string }
        contact_channel: { type: string, enum: [email, sms] }
        contact_value: { type: string }
        host_hint: { type: string, nullable: true }
        privacy_notice_acknowledged: { type: boolean }
        privacy_notice_version: { type: string }
        turnstile_token: { type: string }
    PortalVisitRequestAccepted:
      type: object
      required: [tracking_reference]
      properties:
        tracking_reference: { type: string, pattern: "^REQ-\\d+$" }
...
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }